% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{jonohey_pal}
\alias{jonohey_pal}
\title{Access a jonohey palette}
\usage{
jonohey_pal(
  palette,
  n = NULL,
  type = c("discrete", "continuous"),
  reverse = FALSE
)
}
\arguments{
\item{palette}{palette name}

\item{n}{number of colors. If NULL, returns full palette}

\item{type}{"discrete" or "continuous"}

\item{reverse}{reverse order}
}
\value{
character vector of hex colors, or a function for continuous scales when used by ggplot
}
\description{
Access a jonohey palette
}
