% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jonohey.R
\name{jonohey_save_png}
\alias{jonohey_save_png}
\title{Save a PNG with ragg and transparent device bg}
\usage{
jonohey_save_png(filename, plot, ...)
}
\arguments{
\item{filename}{output filename}

\item{plot}{ggplot object to export with transparent bg}

\item{...}{passed to ggplot2::ggsave()}
}
\description{
Save a PNG with ragg and transparent device bg
}
