% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jonohey_show}
\alias{jonohey_show}
\title{Preview a palette}
\usage{
jonohey_show(palette)
}
\arguments{
\item{palette}{Name of palette to show. Available names can be found using \code{\link[=jonohey_palettes]{jonohey_palettes()}}}
}
\description{
Preview a palette
}
