% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_jonohey}
\alias{scale_color_jonohey}
\title{ggplot2 color scale for jonohey palettes}
\usage{
scale_color_jonohey(palette = "autumn", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{palette name in jonohey_palettes()}

\item{discrete}{logical. If TRUE use discrete scale, else continuous}

\item{reverse}{logical. Reverse palette order}

\item{...}{passed to ggplot2 scale constructors}
}
\description{
Works exactly like \code{\link[=scale_fill_jonohey]{scale_fill_jonohey()}} but applies to \code{color} aesthetics.
}
\seealso{
\code{\link[ggplot2:scale_manual]{ggplot2::scale_color_manual()}}, \code{\link[=scale_fill_jonohey]{scale_fill_jonohey()}}, \code{\link[=jonohey_pal]{jonohey_pal()}}
}
