% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_jonohey}
\alias{scale_fill_jonohey}
\title{ggplot2 scales for jonohey palettes}
\usage{
scale_fill_jonohey(palette = "autumn", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{palette name in jonohey_palettes()}

\item{discrete}{logical. If TRUE use discrete scale, else continuous}

\item{reverse}{logical. Reverse palette order}

\item{...}{passed to ggplot2 scale constructors}
}
\description{
ggplot2 scales for jonohey palettes
}
