% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jonohey.R
\name{theme_axes_wiggle}
\alias{theme_axes_wiggle}
\title{Jonohey axes-wiggle theme}
\usage{
theme_axes_wiggle(
  base_size = 12,
  base_family = "Fuzzy Bubbles",
  amount = 3,
  wiggle_n = 100,
  linewidth = 1,
  line_color = "gray20",
  ...
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{amount}{wiggle intensity (approx. pixels)}

\item{wiggle_n}{number of control points for the wiggle}

\item{linewidth}{linewidth to be passed to axes}

\item{line_color}{color to be passed to axes}

\item{...}{passed to ggplot2 theme constructors}
}
\description{
Only the axes wiggle (cartoony vibe) using elementalist. Geoms remain straight.
No grid, no ticks. Rounded panel optional (same args as round theme).
}
\seealso{
\code{\link[elementalist:wiggle]{elementalist::element_line_wiggle()}}
}
